Move seed groups code to support/groups.rb
